+++ nvision/sim/locs/coarse/sweep_locator.py
@@ -517,7 +517,7 @@ class SweepingLocator(Locator):
         else:
             expected_uniform = float(self.max_steps)
         metrics["expected_uniform_points"] = expected_uniform
-        metrics["measurements_done"] = min(int(round(expected_uniform)), self.max_steps)
+        metrics["measurements_done"] = min(round(expected_uniform), self.max_steps)
         efficiency = expected_uniform / max(metrics["measurements_done"], 1)
         metrics["sweep_efficiency"] = efficiency
         return metrics
